Moved cli.c into examples folder, adjusted Makefile and README
