+++ src/cli/Makefile
@@ -2,9 +2,9 @@
 
 
 CC = gcc
-BINS = cli
-SRCS = src/cmd.c src/shell.c src/cli.c src/parser.c src/operations.c src/validate.c
-HDRS = include/cmd.h include/cli.h include/shell.h include/parser.h include/validate.h
+SRCS = src/cmd.c src/shell.c src/parser.c src/operations.c src/validate.c
+BINS = examples/cli.c
+HDRS = include/cmd.h include/shell.h include/parser.h include/validate.h
 CFLAGS = -fPIC -Wall -Wextra -g -O2 -lm -I./include/ -I../common/include/
 OBJS = $(SRCS:.c=.o)
 LIB = cli.a
@@ -16,7 +16,7 @@ $(LIB): $(SRCS) $(HDRS)
 	$(CC) $(CFLAGS) -shared -o $@ $^
 
 test: $(SRCS) $(HDRS)
-	$(CC) $(CFLAGS) $(SRCS) -o $(@) -lreadline
+	$(CC) $(CFLAGS) $(SRCS) $(BINS) -o $(@) -lreadline
 
 #parser: src/parser.c include/parser.h
 	#$(CC) $(CFLAGS) -o parser parser.c
@@ -24,4 +24,4 @@ test: $(SRCS) $(HDRS)
 .PHONY: clean
 # type 'make clean' to remove the executable and debug information
 clean:
-	rm $(BINS) $(LIB) $(OBJS) test #parser
+	rm $(LIB) test
