+++ src/cli/README
@@ -4,11 +4,6 @@ Besides this README, this directory includes the following files:
 
 - a Makefile for the project
 
-- cli.h, a header file including struct definitions and
-  function prototypes for vcard and bst operations
-
-- cli.c, which runs a shell "chiventure" for an interactive experience
-
 - shell.h and shell.c, shell actions 
 
 - cmd.c and cmd.h that introduces a (to simple) data structure for commands, and navigates between them.
@@ -16,3 +11,6 @@ Besides this README, this directory includes the following files:
 - validate.c and validate.h that check a functions validity
 
 - operations.c and operations.h that will link the operations to actual actions. 
+
+- in the examples folder is a runnable version of the cli, type "make test" in the cli directory
+ in order to compile an executable.
