--- src/cli/include/cli.h
@@ -1,17 +1,0 @@
-#ifndef _CLI_H_
-#define _CLI_H_
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include "cmd.h"
-#include <readline/readline.h>
-#include <readline/history.h>
-
-//HERE WE COULD ADD STRUCTS (SUCH AS STRUCTS FOR DIFFERENT ARGUMENTS)
-
-//As well as headers for different functions we want to call
-
-
-
-#endif /* _CLI_H_ */
